Factored out the core of the LiveServer tests so they can be reused easier
